proba komunkacji z floatem
jak float nie zadziala wracamy do inta dajemy ograniczenia i 3.0 gotowe :)

+++ CM7/Core/Inc/pid.h
@@ -5,17 +5,17 @@
 
 typedef struct
 {
-	int previous_error; 		//Poprzedni błąd dla członu różniczkującego
-	int total_error;		//Suma uchybów dla członu całkującego
+	float previous_error; 		//Poprzedni błąd dla członu różniczkującego
+	float total_error;		//Suma uchybów dla członu całkującego
 	float Kp;			//Wzmocnienie członu proporcjonalnego
 	float Ki;			//Wzmocnienie członu całkującego*/
 	float Kd;			//Wzmocnienie członu różniczkującego*/
-	int anti_windup_limit;		//Anti-Windup - ograniczenie członu całkującego*/
+	float anti_windup_limit;		//Anti-Windup - ograniczenie członu całkującego*/
 }pid_str;
 
-void pid_init(pid_str *pid_data, float kp_init, float ki_init, float kd_init, int anti_windup_limit_init);
+void pid_init(pid_str *pid_data, float kp_init, float ki_init, float kd_init, float anti_windup_limit_init);
 void pid_reset(pid_str *pid_data);
-int pid_calculate(pid_str *pid_data, float setpoint, float process_variable);
+float pid_calculate(pid_str *pid_data, float setpoint, float process_variable);
 
 
 #endif /* INC_PID_H_ */
